refactor(libsy): move escalation router into its own module
