+++ crates/libsy/src/algorithms.rs
@@ -7,6 +7,7 @@
 //! `use switchyard_libsy::Random`.
 
 pub mod advisor_gate;
+mod escalation;
 pub mod fall_through;
 pub mod llm_class;
 pub mod noop;
